refactor: combine modos.introspection with modos.helpers.schema
